requirements_rag.txt
Bio
biopython
boto3
botocore
cellxgene_census
chonkie
fastapi
fastembed
kani
litellm
markitdown
networkx
numpy
openai
pydantic
python-dotenv
qdrant_client
tqdm
uvicorn
